@@ -19,7 +19,7 @@
     "@fastify/pre-commit": "^2.0.2",
     "@sveltejs/adapter-vercel": "^5.1.1",
     "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/vite-plugin-svelte": "^4",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
     "autoprefixer": "^10.4.15",
@@ -33,7 +33,7 @@
     "prettier-plugin-svelte": "^3.0.3",
     "prettier-plugin-tailwindcss": "^0.5.4",
     "rehype-slug": "^6.0.0",
-    "svelte": "^4.2.15",
+    "svelte": "^5",
     "svelte-check": "^3.5.1",
     "svgo": "^3.0.5",
     "tailwindcss": "^3.3.3",
